break out findStaticQueries into util
